Handle overwriting data
The `shmop_write` function does not clear any existing data. If the
original data was longer than the new data, there will be trialing data
after the new data, causing the cache to be corrupt. Adding a null char
to use to mark then end of the string deals with this. If the new string
is longer, it will overwrite the existing null char.